requirements.txt
Flask==3.0.0
requests==2.32.0
jinja2==3.1.4
python-dotenv
urllib3
black==23.3.0
isort==5.12.0
flake8==6.1.0